refactor(auth): remove `ShopifyAuthGuard.register()` requirement
