+++ integration/src/with-hybrid-auth/app.module.ts
@@ -1,4 +1,3 @@
-import { ShopifyAuthModule } from '@nestjs-shopify/auth';
 import { Module } from '@nestjs/common';
 import { ShopifyInitializerModule } from '../shopify-initializer/shopify-initializer.module';
 import { ShopifyOfflineModule } from '../shopify-offline/shopify-offline.module';
@@ -10,7 +9,6 @@ import { HybridController } from './hybrid.controller';
     ShopifyInitializerModule,
     ShopifyOfflineModule,
     ShopifyOnlineModule,
-    ShopifyAuthModule.register(),
   ],
   controllers: [HybridController],
 })
