+++ packages/graphql/src/graphql-proxy/graphql-proxy.module.ts
@@ -1,9 +1,7 @@
-import { ShopifyAuthModule } from '@nestjs-shopify/auth';
 import { Module } from '@nestjs/common';
 import { ShopifyGraphqlProxyController } from './graphql-proxy.controller';
 
 @Module({
-  imports: [ShopifyAuthModule.register()],
   controllers: [ShopifyGraphqlProxyController],
 })
 export class ShopifyGraphqlProxyModule {}
